+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - On Wayland, the window now exists even if nothing has been drawn.
 - On Windows, fix initial dimensions of a fullscreen window.
 - On Windows, Fix transparent borderless windows rendering wrong.
+- On macOS, add `WindowExtMacOS::set_fullscreen_presentation_options` for hiding the dock and the menu bar in fullscreen mode.
 
 # Version 0.19.1 (2019-04-08)